+++ consensus/state_processing/src/common/base.rs
@@ -7,13 +7,13 @@ pub fn get_base_reward<T: EthSpec>(
     state: &BeaconState<T>,
     index: usize,
     // Should be == get_total_active_balance(state, spec)
-    total_active_balance: u64,
+    total_active_balance: u128,
     spec: &ChainSpec,
 ) -> Result<u64, BeaconStateError> {
-    state
-        .get_effective_balance(index)?
-        .safe_mul(spec.base_reward_factor)?
-        .safe_div(total_active_balance.integer_sqrt())?
+    ((state
+        .get_effective_balance(index)? as u128)
+        .safe_mul(spec.base_reward_factor as u128)?
+        .safe_div(total_active_balance.integer_sqrt())? as u64)
         .safe_div(spec.base_rewards_per_epoch)
         .map_err(Into::into)
 }
